+++ .eslintrc
@@ -15,6 +15,8 @@
   "rules": {
     "prettier/prettier": "error",
     "linebreak-style": 0,
-    "arrow-parens": 0
+    "arrow-parens": 0,
+    "implicit-arrow-linebreak": 0,
+    "function-paren-newline": 0
   }
 }
