+++ api/controllers/index.js
@@ -1,0 +1,5 @@
+const AuthController = require('./authController');
+
+module.exports = {
+  AuthController,
+};
