+++ client/src/setupProxy.js
@@ -1,6 +1,7 @@
 const proxy = require('http-proxy-middleware');
 
 module.exports = function(app) {
-  app.use(proxy('/api/*', { target: 'http://[::1]:8080' }));
-  app.use(proxy('/api/auth/*', { target: 'http://[::1]:8080' }));
+  app.use(proxy('/api/*', {
+    target: 'https://[::1]:8080', "secure": false,
+    "changeOrigin": true }));
 };
